requirements.txt
# ─── Core ───
python-dotenv==1.1.0
pyyaml==6.0.2
httpx==0.28.1
tenacity==9.1.2

# ─── Async ───
aiosqlite==0.21.0

# ─── Data & Analysis ───
pandas==2.2.3
pandas-ta==0.3.14b
numpy==2.2.3

# ─── Scheduling ───
apscheduler==3.11.0

# ─── Telegram ───
python-telegram-bot==21.11

# ─── Logging ───
loguru==0.7.3

# ─── LLM ───
anthropic==0.49.0

# ─── Data Providers ───
yfinance==0.2.54
praw==7.8.1
feedparser==6.0.11
gdeltdoc==1.5.0

# ─── India-specific ───
jugaad-data==0.24

# ─── US-specific (install when US system is built) ───
# polygon-api-client==1.14.0
# robin-stocks==3.0.6
